requirements.txt
flask==3.1.2
pypdf==5.9.0
reportlab==4.4.10
Pillow==12.1.1
pdf2image==1.17.0
gunicorn==23.0.0
